+++ src/main/resources/db/changelog/orderingCreateTable.sql
@@ -1,7 +1,7 @@
 create table ordering (
 order_id bigserial primary key,
 order_date date,
-total_cost bigint,
+total_cost FLOAT,
 order_history_id bigint,
 delivery_id bigint REFERENCES delivery(delivery_id)
 )
