+++ src/main/resources/db/changelog/productCreateTable.sql
@@ -2,6 +2,6 @@ create table product (
 product_id bigserial primary key,
 name varchar,
 description varchar,
-price bigint,
+price FLOAT,
 quantity integer
 )
